bugs!

+++ joint_h.m
@@ -2,11 +2,10 @@ function h=joint_h(img1,img2)
 % 
 % calcs the joint histogram of img1 & img2.
 % 
-
 img1 = img1(:);
 img2 = img2(:);
 s = numel(img1);
-N = max(max(img1), max(img2))+1;
+N = double(max(max(img1), max(img2))+1);
 
 h = zeros(N,N);
     for i = 1:s
@@ -14,3 +13,4 @@ h = zeros(N,N);
     end
 
 end
+
